Revert "Cleaned up code"
This reverts commit 6112a707ccde3f038227caf02b97ed13bf3a5952.
